# WIP Re-implement Sets as sorted slice, for performance