--- a/report/string_set.go
+++ b/report/string_set.go
@@ -36,6 +36,22 @@ func (s StringSet) Contains(str string) bool {
 	return i < len(s) && s[i] == str
 }
 
+// ContainsSet returns true if the string set includes all strings in the other set
+func (s StringSet) ContainsSet(b StringSet) bool {
+	i, j := 0, 0
+	for i < len(s) && j < len(b) {
+		if s[i] == b[j] {
+			i++
+			j++
+		} else if s[i] < b[j] {
+			i++
+		} else {
+			return false
+		}
+	}
+	return j == len(b)
+}
+
 // Intersection returns the intersections of a and b
 func (s StringSet) Intersection(b StringSet) StringSet {
 	result, i, j := emptyStringSet, 0, 0