Comparable to #2870 

I should adapt/extend the code-generator from `LatestMap`, really.

Benchmark comparison (the ns timings are extremely variable; I wouldn't read much into them):
```
benchmark                          old ns/op      new ns/op      delta
BenchmarkReportUnmarshal-2         1550377545     1477669912     -4.69%
BenchmarkReportMerge-2             1684092046     1708268415     +1.44%
BenchmarkTopologyList-2            607035752      491860104      -18.97%
BenchmarkTopologyHosts-2           353949174      311919383      -11.87%
BenchmarkTopologyControllers-2     306378851      284710636      -7.07%
BenchmarkTopologyPods-2            240673871      282986959      +17.58%
BenchmarkTopologyContainers-2      188452419      154122803      -18.22%
BenchmarkTopologyProcesses-2       117647086      102960805      -12.48%
BenchmarkSmartMerger-2             78654006       76412340       -2.85%
BenchmarkDumbMerger-2              593382683      429027284      -27.70%

benchmark                          old allocs     new allocs     delta
BenchmarkReportUnmarshal-2         3477338        3137597        -9.77%
BenchmarkReportMerge-2             1951537        1565785        -19.77%
BenchmarkTopologyList-2            821682         810457         -1.37%
BenchmarkTopologyHosts-2           618573         603911         -2.37%
BenchmarkTopologyControllers-2     433439         421355         -2.79%
BenchmarkTopologyPods-2            388042         376551         -2.96%
BenchmarkTopologyContainers-2      261241         251536         -3.71%
BenchmarkTopologyProcesses-2       112362         112382         +0.02%
BenchmarkSmartMerger-2             129514         117970         -8.91%
BenchmarkDumbMerger-2              578302         578325         +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkReportUnmarshal-2         458400096     440474768     -3.91%
BenchmarkReportMerge-2             350652952     351798312     +0.33%
BenchmarkTopologyList-2            92883797      94307848      +1.53%
BenchmarkTopologyHosts-2           65693070      66135388      +0.67%
BenchmarkTopologyControllers-2     46918684      47246880      +0.70%
BenchmarkTopologyPods-2            41667033      42022456      +0.85%
BenchmarkTopologyContainers-2      28547060      28785196      +0.83%
BenchmarkTopologyProcesses-2       14145046      14549248      +2.86%
BenchmarkSmartMerger-2             36853877      35338592      -4.11%
BenchmarkDumbMerger-2              164080260     173093141     +5.49%
```