@rade 2017-12-08
> I should adapt/extend the code-generator from `LatestMap`, really.

Yes, please.

And appease the linter.