@bboreham 2018-05-29
I've actually never stopped working on this; it generated many other PRs in the rendering code as I fought to balance the performance.

I have come to the conclusion that `NodeSet` can't be done efficiently as a slice, and therefore I can't unify all the various `ps.Map`-using code on a new implementation. So we could merge something very close to this PR as a stepping-stone to somewhere less unified.